requirements.txt
azure-core==1.34.0
azure-storage-blob==12.25.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.2
idna==3.10
isodate==0.7.2
numpy==2.2.5
pandas==2.2.3
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
six==1.17.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
